+++ src/components/Report/PdfReports/reportHeaders/StrategyTableHeader.tsx
